RenderInterface: Added ImageFormat opaque handle and ImageLayout enum
These two things describe format of image(e.g. the way its stored in
memory) and what kind of operations they're going to be used for.

Currently ImageFormat is typedef to int64 instead of valid enum. We will
change it when we would need to create images on our own, currently we
only need to get format of already created images(e.g. swapchain)